Extracted LibraryServerTest into a file of its own
